fix(bpf,mount_secure): addr verifier risks and heuristic edge cases
`bpf/headers/module_core.h`
- Update the `BPF_SAFE_MEMSET` macro to use `size_t` for the loop index instead
of a signed integer.

`bpf/mount_secure.bpf.c`:
- Update `is_sensitive_path` to use `len >= 10` to correctly identify exact
`/run/media` mounts without trailing slashes.
- Increase the unrolled loop depth in `fallback_dentry_check` from 4 to 8 to
better catch deeply nested mount paths.
- Check the return values of `bpf_probe_read_kernel_str` and apply an "N/A"
fallback if string extraction fails.
- Guard `container_of` pointer arithmetic with `bpf_core_field_exists` to
prevent load-time verifier rejections on unsupported kernel BTFs, falling back
to the superblock ID.

+++ bpf/headers/module_core.h
@@ -52,7 +52,7 @@
 #define BPF_SAFE_MEMSET(dest, size)                                            \
 	do {                                                                       \
 		volatile __u8 *__ptr = (volatile __u8 *)(dest);                        \
-		for (int __i = 0; __i < (size); __i++) {                               \
+		for (size_t __i = 0; __i < (size_t)(size); __i++) {                    \
 			__ptr[__i] = 0;                                                    \
 		}                                                                      \
 	} while (0)
